requirements/test-ci.txt
pytest
pytest-django
coverage
